- Fixing up some compile warnings
git-svn-id: svn://70.113.44.214/VanaMirror/branches/buis_vana_overhaul@3266 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/Common/ConfigFile.cpp
@@ -139,7 +139,7 @@ vector<int8_t> ConfigFile::getBossChannels(const string &value, size_t maxChanne
 	lua_getglobal(getLuaState(), value.c_str());
 	lua_pushnil(getLuaState());
 	while (lua_next(getLuaState(), -2)) {
-		channels.push_back(lua_tointeger(getLuaState(), -1));
+		channels.push_back(static_cast<int8_t>(lua_tointeger(getLuaState(), -1)));
 		lua_pop(getLuaState(), 1);
 	}
 	lua_pop(getLuaState(), 1);
